fix: separate server/client anonymous session utilities to resolve build error
The build was failing because Next.js `cookies()` import was being used in
a client component. This change separates the server-side and client-side
utilities into different files to resolve the webpack compilation error.

## Changes Made

- **New file**: `lib/auth/anonymous-session-server.ts` - Server-side utilities
  - Contains `getAnonymousSessionIdFromCookie()`, `setAnonymousSessionIdCookie()`
  - Uses `next/headers` cookies import (server-only)

- **Modified**: `lib/auth/anonymous-session.ts` - Client-side utilities only
  - Removed server-side functions that used `next/headers`
  - Kept all client-side functions for browser environment

- **Updated API routes**: Changed imports to use server-side utilities
  - `app/api/diagnostic/route.ts`
  - `app/api/diagnostic/session/[id]/status/route.ts`

## Resolution
- Build now succeeds without webpack errors
- Client components can import client-side utilities safely
- Server components/API routes use server-side utilities
- All tests continue to pass
- Maintains same functionality with proper separation of concerns
