+++ app/api/diagnostic/session/[id]/status/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { createServerSupabaseClient } from '@/lib/supabase/server';
-import { getAnonymousSessionIdFromCookie } from '@/lib/auth/anonymous-session';
+import { getAnonymousSessionIdFromCookie } from '@/lib/auth/anonymous-session-server';
 
 export async function GET(req: Request) {
   const supabase = createServerSupabaseClient();
